Prod dep features (#84)
* Added button Reklamacja

* Added btnReturn

* Added searching complaints

* Added FormSalesDepartmentComplaintAdd and modified data in dql,dml.

* Button was modified in FormSalesDepartmentComplaintAdd.

* FormNewOrder and FormNewOrderDetails were added. Files ddl, dql, dml were corrected.

* Adding standard order was corrected.

* CommertialOffer and NewOrder were added.

* Problem with new order and new details are fixed. Two views are added.

* FormNewCommertialOffer.cs was corrected

* FormNewCommertialOffer.cs was added.

* FormSalesDepartmentOffersForConsideration was added.

* Problems with designer forms were fixed. MessageBox in FormSalesDepartmentSales was fixed. FormSalesDepartmentOffersForConsideration working was fixed.

* Switch case was added

* Added "FormSalesDepartmentOrders" and "FormSalesDepartmentOrderDetails"

* Sales department works completely.

* MaximizeBox in FormSalesDepartmentAddSalesDetails was corrected.

* ddl and FormSalesDepartmentAddSalesDetails were corrected.

* dql was corrected.

* MessageBox was added

* FormSalesDepartmentComplaintAdd was corrected.

* dql was corrected and conflicts were resolved.

* Resolve issues.

* Add modyfication to KontrolaJakosci Form

* Fixed issues except var type. Time in Form1 was fixed.

* FormSalesDepartmentComplaintAdd was fixed.

* Najs czendżys

* Removing cast

* Adding generate button to form Wytwarzanie

* Generowanie danych w dziale kontroli + dodatkowe widoki dql.sql

* Adding generator for quiality control

* Charts, views in dql were added. Cast, var type were deleted.

* Conflicts resolved.

* Change Form Kontrola Jakosci

* Removing bugs

* Branch able to merge

* Resolve issuess SalesDep

* Revert "Resolve issuess SalesDep"

This reverts commit 6a70ef6173a1792aaedf54c9a7221b3069458b71.

* Resolve issuess SalesDep

* Able to marge

+++ KWZP2022/KWZP2022/v_Sprzedaz.cs
@@ -21,6 +21,6 @@ namespace KWZP2022
         public System.DateTime Data_początku_sprzedaży { get; set; }
         public System.DateTime Data_końca_sprzedaży { get; set; }
         public int Umowa { get; set; }
-        public Nullable<decimal> Koszt { get; set; }
+        public Nullable<int> Koszt { get; set; }
     }
 }
